pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Shazil <shazilva2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.4.0"
djangorestframework-simplejwt = "^5.3.1"
django-phonenumber-field = "^8.0.0"
psycopg2 = "^2.9.9"
phonenumbers = "^8.13.40"
pillow = "^10.4.0"
python-decouple = "^3.8"
pytest-django = "^4.8.0"
cryptography = "^43.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
